MOB-1741 - Use RK properly to load the dashboards from server. Fix outlet link in Gadget display view xib. Prevents from reloading the same view from the menu, which prevents crashes when the same proxy is loaded over and over again.

+++ Sources/iPad/View Controllers/CascadeViewController/MenuViewController.h
@@ -21,14 +21,14 @@
                                                   SettingsDelegateProcotol, AccountSwitcherDelegate> {
                                                       
     
-	UITableView*  _tableView;
+	UITableView*    _tableView;
 	NSMutableArray* _cellContents;
     
-	int									_intSelectedLanguage;
+	int				_intSelectedLanguage;
 
-    int                             _intIndex;
+    NSInteger       _intIndex;
     
-    BOOL                            _isCompatibleWithSocial;
+    BOOL            _isCompatibleWithSocial;
     
 }
 @property (nonatomic, retain) UITableView* tableView;
